fixed application/json checking changed getSetValue to be basic_device example
